+++ src/faucet/mainnet/views.rs
@@ -1,5 +1,5 @@
 use crate::faucet::mainnet::{FIL_MAINNET_UNIT, MAINNET_DRIP_AMOUNT, MAINNET_RATE_LIMIT_SECONDS};
-use crate::faucet::views::Faucet;
+use crate::faucet::views::faucet::Faucet;
 use crate::rpc_context::{Provider, RpcContext};
 use crate::utils::format::format_balance;
 use fvm_shared::address::Network;
