+++ src/faucet/views/mod.rs
@@ -1,0 +1,8 @@
+pub mod alert;
+pub mod balance;
+pub mod faucet;
+pub mod home;
+pub mod icons;
+pub mod layout;
+pub mod nav;
+pub mod transaction;
