+++ src/utils/mod.rs
@@ -1,6 +1,5 @@
 pub mod address;
 pub mod error;
 pub mod format;
-pub mod icons;
 pub mod key;
 pub mod message;
